chore: fix unnecessary dependency of tx settlement on offchain aggregator
