+++ packages/valory/agents/oracle/aea-config.yaml
@@ -31,13 +31,13 @@ protocols:
 skills:
 - valory/abstract_abci:0.1.0:bafybeidt4cyfhlla6faca6us6law7dijrlrzihcj6uwrmeaa3ho6fdck2u
 - valory/abstract_round_abci:0.1.0:bafybeigpz64qx22a4zzlcczmc5frhlecytrh5q3x77lwe3a5kee2qrcchm
-- valory/oracle_abci:0.1.0:bafybeib5vhkaltdem6z2ybzdevqiuepsnndtgilefwtz2lydhfv6cc36he
+- valory/oracle_abci:0.1.0:bafybeidggop5kc4b65qxtvdyzsmohpv6flqxr2ybxfglngbmcjve76ggym
 - valory/oracle_deployment_abci:0.1.0:bafybeigadpn6vplc2hnyahsgpxndzqrkgqnxjenbufojyqy6py7t4l2eya
-- valory/price_estimation_abci:0.1.0:bafybeifkflud5zv6qajiwzsp56hhtlfa3jjfj2pcehzgglakna3bwvfh7e
+- valory/price_estimation_abci:0.1.0:bafybeigthpgeb6jgxz3agymonbrgxn3rqjzdunuklj5mmoz3v4437imdfm
 - valory/registration_abci:0.1.0:bafybeiajvfkffbdmggj3it7e4r6733lylxhbnb2a6tx3tmlgpy2fpxkuze
 - valory/reset_pause_abci:0.1.0:bafybeieo5m37rwoel7e6u3lrp2cei7hdyhklggsc6776vej6bp6gzua7ei
 - valory/safe_deployment_abci:0.1.0:bafybeifkxk6jkixnk4fymsw76zw2z2ofshmomqbxkp2kcwpur6dn7jb44i
-- valory/transaction_settlement_abci:0.1.0:bafybeig3lpuertmg5wngcnherollnvakvqgeisy3a6f5tbmw4e5k5khnt4
+- valory/transaction_settlement_abci:0.1.0:bafybeifazkkkyqn6pjxl5jnl3qmuv75i4lnjssww2aimjnjy3qcaatqsry
 default_ledger: ethereum
 required_ledgers:
 - ethereum
