+++ packages/valory/services/oracle_hardhat/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeieqpbuzrxtfo5hz3gsb64aumnfbxad4tpxp4s46jxj7byb3zxx34y
 fingerprint_ignore_patterns: []
-agent: valory/oracle:0.1.0:bafybeiarh3wkocultmi3i6athorlwxj2hrs2toife7lxcerhf6kakvyoea
+agent: valory/oracle:0.1.0:bafybeihcomi2hqq4l3nkrjtzkv3npni6zyaqhfs2bwkvwh3azrapvgalom
 number_of_agents: 4
 ---
 setup_args:
