+++ packages/valory/skills/liquidity_rebalancing_abci/skill.yaml
@@ -36,7 +36,7 @@ protocols:
 - valory/contract_api:1.0.0:bafybeidyjrthhq3p27k3v3a4xx7b7lu4fe4h765gmkyyyj7xur4b25bxim
 skills:
 - valory/abstract_round_abci:0.1.0:bafybeigpz64qx22a4zzlcczmc5frhlecytrh5q3x77lwe3a5kee2qrcchm
-- valory/transaction_settlement_abci:0.1.0:bafybeig3lpuertmg5wngcnherollnvakvqgeisy3a6f5tbmw4e5k5khnt4
+- valory/transaction_settlement_abci:0.1.0:bafybeifazkkkyqn6pjxl5jnl3qmuv75i4lnjssww2aimjnjy3qcaatqsry
 behaviours:
   main:
     args: {}
