+++ packages/valory/skills/oracle_abci/tests/test_integration.py
@@ -67,6 +67,7 @@ from packages.valory.skills.price_estimation_abci.behaviours import (
 from packages.valory.skills.price_estimation_abci.rounds import (
     SynchronizedData as PriceEstimationSynchronizedSata,
 )
+from packages.valory.skills.price_estimation_abci.tests import GnosisIntegrationBaseCase
 from packages.valory.skills.transaction_settlement_abci.behaviours import (
     CheckLateTxHashesBehaviour,
     FinalizeBehaviour,
@@ -81,9 +82,6 @@ from packages.valory.skills.transaction_settlement_abci.payload_tools import (
 )
 from packages.valory.skills.transaction_settlement_abci.rounds import (
     SynchronizedData as TxSettlementSynchronizedSata,
-)
-from packages.valory.skills.transaction_settlement_abci.test_tools.integration import (
-    GnosisIntegrationBaseCase,
 )
