+++ packages/valory/skills/transaction_settlement_abci/test_tools/integration.py
@@ -29,16 +29,12 @@ from typing import Any, Dict, cast
 from aea.crypto.base import Crypto
 from aea.crypto.registries import make_crypto, make_ledger_api
 from aea_ledger_ethereum import EthereumApi
-from aea_test_autonomy.fixture_helpers import HardHatAMMBaseTest
 from aea_test_autonomy.helpers.contracts import get_register_contract
 from web3.types import Nonce, Wei
 
 from packages.open_aea.protocols.signing import SigningMessage
 from packages.valory.contracts.gnosis_safe.tests.test_contract import (
     PACKAGE_DIR as GNOSIS_SAFE_PACKAGE,
-)
-from packages.valory.contracts.offchain_aggregator.tests.test_contract import (
-    PACKAGE_DIR as OFFCHAIN_AGGREGATOR_PACKAGE,
 )
 from packages.valory.protocols.contract_api import ContractApiMessage
 from packages.valory.protocols.contract_api.custom_types import RawTransaction, State
@@ -53,7 +49,6 @@ from packages.valory.skills.abstract_round_abci.test_tools.integration import (
     ExpectedTypesType,
     HandlersType,
     IntegrationBaseCase,
-    _HarHatHelperIntegration,
 )
 from packages.valory.skills.transaction_settlement_abci.behaviours import (
     FinalizeBehaviour,
@@ -317,19 +312,3 @@ class _TxHelperIntegration(_GnosisHelperIntegration):
                 final_verification_status=VerificationStatus.VERIFIED,
                 final_tx_hash=self.tx_settlement_synchronized_data.to_be_validated_tx_hash,
             )
-
-
-class GnosisIntegrationBaseCase(
-    _TxHelperIntegration, _HarHatHelperIntegration, HardHatAMMBaseTest
-):
-    """Base test class for integration tests in a Hardhat environment, with Gnosis deployed."""
-
-    # TODO change this class to use the `HardHatGnosisBaseTest` instead of `HardHatAMMBaseTest`.
-
-    @classmethod
-    def setup(cls, **kwargs: Any) -> None:
-        """Setup."""
-        super().setup()
-
-        # register offchain aggregator contract
-        _ = get_register_contract(OFFCHAIN_AGGREGATOR_PACKAGE)
